backend/requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.8.0
scikit-learn>=1.3.2
numpy>=1.26.2
mlflow>=2.9.2
python-dotenv==1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
qdrant-client>=1.7.0
openai>=1.0.0
sentence-transformers>=2.2.0


